Rename Main2 into Main
(delete old Main)
